Update Part 1 - Introduction to Kalman Filter.ipynb
Managed to code up the 1D kalman filter and plot out the results. So far it looks similar to what is shown in the reference text by Alex Becker.
It would be wise to consider adding lean and sharp markdown notes to complement the code.
So far i am liking this code heavy notebook in contrast to the spacecraft dynamics and control repo!