Update Kalman_Filter_class_development.ipynb
Did a multivariate LinearKF example with reference to Alex Becker's text.
The system and framework in developing Kalman Filters and its various flavors seem promising, but consider the following improvements:
- Logging specific parameters? Or ways in which we can experiment and shift what needs to be logged? At the moment, only the developer of this code (me) would know what is logging. Obviously more data is good, so there could be further considerations on how to make KalmanLogger more robust.
- Review every single code and begin implementing pytest, unit test every class and functionality. It would be prudent to also test for error-catching to find blindspots in what has been developed so far.
- Code ExtendedKF
- Code UnscentedKF

It might be wise to write a technical document detailing considerations and design of the code being developed here. It could potentially spin-off into a Kalman Filter sandbox, KalmanBox? This could be a way to rapidly prototype various kalman filters found in several literature.